bhavalay_backend/consult_now/serializers.py
from rest_framework import serializers
from .models import Consultation

class ConsultationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Consultation
        fields = ['name', 'age', 'occupation', 'course', 'mobile_number', 'email']