bhavalay_backend/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
idna==3.4
mysql-connector==2.2.9
Pillow==9.5.0
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
